refactor(task): 拆分 TaskService 并重构任务生命周期
- 将 task 公共协议/上下文/异常/schema 下沉到 core，并迁移 polling 目录
- 新增 execute/submit/video 子模块，拆分同步执行、异步提交流程、错误处理与视频任务操作
- 收敛 TaskService 为门面编排，内部委派到 SyncTaskExecutionService、AsyncTaskSubmitService、VideoTaskOperationsService
- 重构 main 生命周期管理：引入 LifecycleState，拆分启动与关闭流程
- 更新并补充测试，覆盖 submit 分层委派、状态流转、任务池操作与视频取消委派
