+++ src/api/handlers/base/chat_sync_executor.py
@@ -160,7 +160,7 @@ class ChatSyncExecutor:
 
             # 统一入口：总是通过 TaskService
             from src.services.task import TaskService
-            from src.services.task.context import TaskMode
+            from src.services.task.core.context import TaskMode
 
             exec_result = await TaskService(handler.db, handler.redis).execute(
                 task_type="chat",
