+++ src/api/handlers/base/cli_stream_mixin.py
@@ -162,7 +162,7 @@ class CliStreamMixin:
 
             # 统一入口：总是通过 TaskService
             from src.services.task import TaskService
-            from src.services.task.context import TaskMode
+            from src.services.task.core.context import TaskMode
 
             exec_result = await TaskService(self.db, self.redis).execute(
                 task_type="cli",
