+++ src/services/candidate/failover.py
@@ -16,9 +16,9 @@ from src.models.database import RequestCandidate
 from src.services.orchestration.error_classifier import ErrorAction, ErrorClassifier
 from src.services.request.candidate import RequestCandidateService
 from src.services.scheduling.schemas import PoolCandidate, ProviderCandidate
-from src.services.task.exceptions import StreamProbeError
-from src.services.task.protocol import AttemptFunc, AttemptKind, AttemptResult
-from src.services.task.schema import ExecutionResult
+from src.services.task.core.exceptions import StreamProbeError
+from src.services.task.core.protocol import AttemptFunc, AttemptKind, AttemptResult
+from src.services.task.core.schema import ExecutionResult
 
 from .policy import FailoverAction, RetryMode, RetryPolicy, SkipPolicy
 from .recorder import CandidateRecorder
