--- src/services/task/impl/__init__.py
@@ -1,7 +1,0 @@
-"""Per-task-type implementations (Phase2).
-
-Currently includes:
-- video: polling adapter
-"""
-
-__all__ = []
